% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AKmoose}
\alias{AKmoose}
\title{Data Set with Alaska Moose Observations.}
\format{
A spatial polygons (\code{sp}) object inclding:
\describe{
  \item{CENTRLAT}{The latitude of the centroid for each site}
  \item{CENTRLON}{The latitude of the centroid for each site}
  \item{STRAT}{A stratification variable}
  \item{TOTAL}{The total moose count on each site}
  ...
}
}
\usage{
AKmoose
}
\description{
A data set that can be used with the \code{sptotal} package. In
this example, the counts are of moose on 860 sites of equal area.
}
\keyword{datasets}
