% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.slmfit.R
\name{check.variogram}
\alias{check.variogram}
\title{Plot an Empirical Variogram of Residuals}
\usage{
check.variogram(object)
}
\arguments{
\item{object}{is an object of class \code{slmfit}.}
}
\value{
a plot of the empirical variogram with the fitted model overlayed.
}
\description{
Plots an empirical variogram of the residuals from the spatial linear model
with the fitted parametric model as a curve overtop of the
binned points.
}
