% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.slmfit.R
\name{coef.slmfit}
\alias{coef.slmfit}
\title{Extract Model Coefficients}
\usage{
\method{coef}{slmfit}(object, ...)
}
\arguments{
\item{object}{a slmfit object}

\item{...}{additional arguments}
}
\value{
a vector of fitted model coefficients.
}
\description{
Extract Model Coefficients
}
