% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corModels.R
\name{corModelExponential}
\alias{corModelExponential}
\alias{corModelGaussian}
\alias{corModelCauchy}
\alias{corModelSpherical}
\alias{corModelCubic}
\title{Spatial Correlation Models}
\usage{
corModelExponential(distance.matrix, range)

corModelGaussian(distance.matrix, range)

corModelCauchy(distance.matrix, range, extrap)

corModelSpherical(distance.matrix, range)

corModelCubic(distance.matrix)
}
\arguments{
\item{distance.matrix}{The distance matrix for sampled sites}

\item{range}{The range that determines how quickly covariance
among sites tapers}

\item{extrap}{Smoothness parameter (optional)}
}
\value{
Correlation Matrix
}
\description{
Spatial Correlation Models
}
\section{Functions}{
\itemize{
\item \code{corModelGaussian}: Gaussian Correlation Structure

\item \code{corModelCauchy}: Cauchy Correlation Structure

\item \code{corModelSpherical}: Spherical Correlation Structure

\item \code{corModelCubic}: Cubic Correlation Structure
}}

