% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estcovparm_jay.R
\name{estcovparm_jay}
\alias{estcovparm_jay}
\title{Estimate Covariance Parameters}
\usage{
estcovparm_jay(response, designmatrix, xcoordsvec, ycoordsvec,
  CorModel = "Exponential", estmethod = "REML")
}
\arguments{
\item{response}{a vector of a response variable, possibly with
missing values.}

\item{designmatrix}{is the matrix of covariates used to regress
the response on.}

\item{xcoordsvec}{is a vector of x coordinates}

\item{ycoordsvec}{is a vector of y coordinates}

\item{CorModel}{is the covariance structure. By default, \code{covstruct} is}

\item{estmethod}{is either the default \code{"REML"} for restricted
maximum likelihood to estimate the covariance parameters and
regression coefficients or \code{"ML"} to estimate the covariance
parameters and regression coefficients.
Exponential but other options include the Spherical and the Gaussian.}
}
\value{
a list with \itemize{
   \item a vector of estimated covariance parameters
   \item the fitted covariance matrix for all of the sites
}
}
\description{
Used to estimate spatial covariance parameters for a few different spatial models.
Estimated parameters can then be used in \code{FPBKpred} to predict unobserved values.
}
\details{
The function is used internally in \code{FPBKpred}.
}
