% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPBKForestPack.R
\docType{package}
\name{FPBKForestPack}
\alias{FPBKForestPack}
\alias{FPBKForestPack-package}
\title{ForestPack: A package used for performing finite population block
kriging on polygonal count data.}
\description{
The package provides an option to perform FPBK on counts assuming
perfect detection.
}
\details{
\code{FPBKForestPack} Main Functions:

\code{FPBKpred} is the function used to generate the kriged predictions
on the unsampled sites.

\code{FPBKoutput} takes the resulting object from FPBKpred to construct
confidence intervals, variograms, and maps.
}
