% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.sptotalPredOut.r
\name{get.predplot}
\alias{get.predplot}
\title{Creates a default map from predictions}
\usage{
get.predplot(x)
}
\arguments{
\item{x}{the output of the prediction function, of class sptotalPredOut}
}
\value{
a plot with predictions, with points filled by whether or not a site was sampled.
}
\description{
Creates a default map from predictions. Note that all predictions are stored
in a data frame in the output of \code{predict}. Therefore, if a user
would like to create his or her own plot, he or she can easily do so using
this data frame.
}
