% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GR2.R
\name{GR2}
\alias{GR2}
\title{Computes the Generalized R-squared.}
\usage{
GR2(object)
}
\arguments{
\item{object}{is an object of class \code{slmfit}.}
}
\value{
the value for generalized R-squared
}
\description{
Computes the Generalized R-squared.
}
