% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.slmfit.R, R/logLik.slmfit.R
\name{logLik.slmfit}
\alias{logLik.slmfit}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{slmfit}(object)

\method{logLik}{slmfit}(object)
}
\arguments{
\item{object}{is a slmfit object}

\item{object}{is a slmfit object}
}
\value{
the log-likelihood of the slmfit object

the log-likelihood of the slmfit object
}
\description{
Extract Log-Likelihood

Extract Log-Likelihood
}
