% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mginv.R
\name{mginv}
\alias{mginv}
\title{Constructing the generalized inverse of a matrix}
\usage{
mginv(X, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{The matrix to be inverted}

\item{tol}{The tolerance of the estimation}
}
\value{
The generalized inverse matrix
}
\description{
Computes the generalized inverse of a matrix X
}
\details{
This function is used in the \code{m2LL.FPBK.nodet} functions in order
to estimate the spatial covariance parameters
}
