% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ORbiomass}
\alias{ORbiomass}
\title{Oregon Biomass Data}
\format{
A data frame object inclding:
\describe{
  \item{coordinates}{The spatial cooredinates for the centroid for each site}
  \item{BIOT_ha}{Biomass on Each Site}
  \item{precip_mean_30yr}{Average precipitation for 30 years}
  \item{temp_mean_30yer}{Average temperature for 30 years}
  ...
}
}
\usage{
ORbiomass
}
\description{
A data set that can be used with the \code{sptotal} package. In
this example, the response variable is a density.
}
\keyword{datasets}
