% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointSimCSR.R
\name{pointSimCSR}
\alias{pointSimCSR}
\title{simulate completely spatially random point patterns.}
\usage{
pointSimCSR(npoints = 100, lower_x_lim = 0, upper_x_lim = 1,
  lower_y_lim = 0, upper_y_lim = 1)
}
\arguments{
\item{npoints}{number of points to add that are completely spatially random (CSR), default = 100}

\item{lower_x_lim}{left limit of boundary, default = 0}

\item{upper_x_lim}{right limit of boundary, default = 1}

\item{lower_y_lim}{lower limit of boundary, default = 0}

\item{upper_y_lim}{upper limit of boundary, default = 1}
}
\value{
data.frame of two columns, x-coordinate in the first, and y-coordinate in the second.
}
\description{
simulates a completely spatially random point patterns. This function is
only used in simulating data sets.
}
\author{
Jay Ver Hoef
}
