% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointSimSyst.R
\name{pointSimSyst}
\alias{pointSimSyst}
\title{Creates a systematic grid of points.}
\usage{
pointSimSyst(
  nrow = 10,
  ncol = 10,
  lower_x_lim = 0,
  upper_x_lim = 1,
  lower_y_lim = 0,
  upper_y_lim = 1
)
}
\arguments{
\item{nrow}{the number of rows in the systematic grid, default is 10}

\item{ncol}{the number of cols in the systematic grid, default is 10}

\item{lower_x_lim}{the lower limit for x-coordinate, default is 0}

\item{upper_x_lim}{the upper limit for x-coordinate, default is 1}

\item{lower_y_lim}{the lower limit for y-coordinate, default is 0}

\item{upper_y_lim}{the upper limit for y-coordinate, default is 1}
}
\value{
A data.frame with x- and y-coordinates of simulated locations
}
\description{
Creates a systematic grid of points. This function is only used
in simulating data sets.
}
\author{
Jay Ver Hoef
}
