% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.slmfit.R
\name{predict.slmfit}
\alias{predict.slmfit}
\title{Perform Finite Population Block Kriging}
\usage{
\method{predict}{slmfit}(object, wtscol = NULL, ...)
}
\arguments{
\item{object}{is an object generated from \code{\link{slmfit}()}}

\item{wtscol}{is the name of the column that contains the weights for prediction.}

\item{...}{further arguments passed to or from other methods.
The default setting predicts the population total}
}
\value{
a list with \itemize{
  \item the estimated population total
  \item the estimated prediction variance
  \item a data frame containing \enumerate{
       \item x-coordinates
       \item y-coordinates
       \item density predictions
       \item count predictions
       \item site-by-site density prediction variances
       \item site-by-site count prediction variances
       \item indicator variable for whether or not the each site was sampled
       \item estimated mean for each site
       \item area of each site
       }
   \item vector with estimated covariance parameters
   \item the formula used to fit the model in \code{slmfit()}
}
}
\description{
Uses an object generated from \code{\link{slmfit}()} to predict the counts on the unsampled sites.
The column of the data set that has the counts should have numeric values for the observed counts
on the sampled sites and `NA` for any site that was not sampled.
}
\examples{
data(exampledataset) ## load a toy data set
slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
predict(slmobj)
}
