% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.slmfit.R
\name{residuals.slmfit}
\alias{residuals.slmfit}
\title{Extract Model Residuals}
\usage{
\method{residuals}{slmfit}(object, ...)
}
\arguments{
\item{object}{a slmfit object}

\item{...}{additional arguments}
}
\value{
a vector of residuals, consisting of each observed count minus the estimated mean.
}
\description{
Extract Model Residuals
}
