% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated Spatially Autocorrelated Data.}
\format{
A data frame object inclding:
\describe{
  \item{x}{The x-coordinate of the centroid for each site}
  \item{y}{The y-coordinate of the centroid for each site}
  \item{X1}{Independent Variable to be used as a predictor}
  \item{X2}{Independent Variable to be used as a predictor}
  \item{X3}{Independent Variable to be used as a predictor}
  \item{X4}{Independent Variable to be used as a predictor}
  \item{X5}{Independent Variable to be used as a predictor}
  \item{X6}{Spatially Correlated Random Variable}
  \item{X7}{Spatially Correlated Random Variable}
  \item{F1}{Factor Variable to be used as a predictor}
  \item{F2}{Factor Variable to be used as a predictor}
  \item{Z}{The total moose count on each site}
  \item{wts1}{Possible Weighting Column to be used as the prediction weights}
  \item{wts2}{Possible Weighting Column to be used as the prediction weights}
  ...
}
}
\usage{
simdata
}
\description{
A data set that can be used with the \code{sptotal} package. In
this example, the response variable is a density.
}
\keyword{datasets}
