% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slmfit.R
\name{slmfit}
\alias{slmfit}
\title{Fits a Spatial Linear Model}
\usage{
slmfit(formula, data, xcoordcol, ycoordcol, CorModel = "Exponential",
  coordtype = "LatLon")
}
\arguments{
\item{formula}{is an R linear model formula specifying density as the
response variable as well as covariates for predicting densities on the unsampled sites.}

\item{data}{is the data set with the response column of densities, the covariates to
be used for the block kriging, and the spatial coordinates for all of the sites.}

\item{xcoordcol}{is the name of the column in the data frame with x coordinates or longitudinal coordinates}

\item{ycoordcol}{is the name of the column in the data frame with y coordinates or latitudinal coordinates}

\item{CorModel}{is the covariance structure. By default, \code{CorModel} is
Exponential but other options include the Spherical and Gaussian.}

\item{coordtype}{specifies whether spatial coordinates are in latitude, longitude (\code{LatLon}) form or UTM (\code{UTM}) form.}
}
\value{
a list with \itemize{
  \item the spatial covariance estimates
  \item the regression coefficient estimates
  \item a list containing \enumerate{
       \item formula
       \item data
       \item xcoordcol
       \item ycoordcol
       \item CorModel
       \item Inverted covariance matrix on the sampled sites
       \item Covariance matrix on all sites
       }
}
}
\description{
Estimates regression coefficients and spatial autocorrelation
parameters.
}
