% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptotal.R
\docType{package}
\name{sptotal}
\alias{sptotal}
\alias{sptotal-package}
\title{sptotal: A package used for performing finite population block
kriging on polygonal count data.}
\description{
The package provides an option to perform FPBK on counts assuming
perfect detection of counts on the sites that were in the survey sample.
}
\details{
\code{sptotal} Main Functions:

\code{slmfit} fits a spatial linear model to the response on the
observed/sampled sites. \code{check.variogram} can be used to construct
an empirical variogram of the residuals of the spatial linear model.

\code{FPBKpred} uses the spatial linear model in \code{slmfit} and finite
population block kriging to predict counts/densities at unobserved locations.
A prediction for the total count as well as a prediction variance
are given by default.

\code{get.predinfo} and \code{get.predplot} take the resulting object from
 \code{FPBKpred} to construct (1) summary information, including the
 prediction, prediction variance, and a prediction interval as well as
 (2) a plot of the site-wise predictions.

Most of the remaining functions in the package are either helper functions
or extra optional functions to extract various specific things from an
\code{slmfit} object, such as residuals, AIC, log-likelihood, etc.
}
