% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USlakes}
\alias{USlakes}
\title{U.S. Lakes Data}
\format{A data frame object inclding:
\describe{
  \item{XCOORD}{The x spatial cooredinate for the centroid for each site}
  \item{YCOORD}{The y spatial cooredinate for the centroid for each site}
  \item{DOC_RESULT}{Disolved Organic Carbon}
  \item{ELEVATION}{Variable that is a candidate for prediction}
  \item{FCIBIG_LIT}{Variable that is a candidate for prediction}
  \item{RVFCGNDBARE_RIP}{Variable that is a candidate for prediction}
  \item{RVFCGNDWOODY_RIP}{Variable that is a candidate for prediction}
  \item{RVFPUNDWOODY_RIP}{Variable that is a candidate for prediction}
  \item{UID}{Lake Identification Number}
  ...
}}
\usage{
USlakes
}
\description{
A data set that can be used with the \code{sptotal} package. In
this example, the lakes are located across the United States.
}
\keyword{datasets}
