% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.slmfit.R
\name{AIC.slmfit}
\alias{AIC.slmfit}
\title{Extract the AIC from a \code{slmfit} object for comparing models.}
\usage{
\method{AIC}{slmfit}(object, ...)
}
\arguments{
\item{object}{a \code{slmfit} object}

\item{...}{additional arguments}
}
\value{
The AIC value of the slmfit object.  Here, AIC is computed as 2 times the
negative log-likelihood plus 2 times the number of model parameters.  Thus,
lower AIC values are preferred models.
}
\description{
Extract the AIC from a \code{slmfit} object for comparing models.
}
\examples{
data(exampledataset) ## load a toy data set
slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
AIC(slmobj)
}
