% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.variogram.R
\name{check.variogram}
\alias{check.variogram}
\title{Plot an Empirical Variogram of Residuals}
\usage{
check.variogram(object)
}
\arguments{
\item{object}{is an object of class \code{slmfit}.}
}
\value{
a plot of the empirical variogram with the fitted model overlayed.
}
\description{
Plots an empirical variogram of the residuals from the spatial linear model
with the fitted parametric model as a curve overtop of the
binned points. By default, the empirical variogram only shows distances
that are less than or equal to the maximum distance in the data
set divided by 2. Therefore, it's possible that the REML-fitted model
will not "fit" the points perfectly.
}
