% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corModelExponential}
\alias{corModelExponential}
\alias{corModelGaussian}
\alias{corModelSpherical}
\title{Spatial Correlation Models}
\usage{
corModelExponential(distance.matrix, range)

corModelGaussian(distance.matrix, range)

corModelSpherical(distance.matrix, range)
}
\arguments{
\item{distance.matrix}{The distance matrix for sampled sites}

\item{range}{The range that determines how quickly covariance
among sites tapers}
}
\value{
Correlation Matrix
}
\description{
Note that, currently, only three of these models are implemented
in the \code{sptotal} package: \code{corModelExponential()},
 \code{corModelGaussian()}, and \code{corModelSpherical()}.
}
\section{Functions}{
\itemize{
\item \code{corModelExponential()}: Exponential Correlation Structure

\item \code{corModelGaussian()}: Gaussian Correlation Structure

\item \code{corModelSpherical()}: Spherical Correlation Structure

}}
