% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPBKpred.R
\name{FPBKpred}
\alias{FPBKpred}
\title{Perform Finite Population Block Kriging}
\usage{
FPBKpred(slmfitobj, FPBKcol = NULL)
}
\arguments{
\item{slmfitobj}{is an object generated from \code{slmfit}}

\item{FPBKcol}{is the name of the column that contains the weights for
prediction. The default setting predicts the population total}
}
\value{
a list with \itemize{
  \item the estimated population total
  \item the estimated prediction variance
  \item a data frame containing \enumerate{
       \item x-coordinates
       \item y-coordinates
       \item density predictions
       \item indicator variable for whether or not the each site was sampled
       \item estimated mean for each site
       }
   \item vector with estimated covariance parameters
}
}
\description{
Takes the output of \code{slmfit} and uses FPBK to predict the counts on the unsampled sites.
The column with the counts should have numeric values for the observed counts
on the sampled sites and `NA` for any site that was not sampled.
}
