% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.predinfo.R
\name{get.predinfo}
\alias{get.predinfo}
\title{Display basic summary information in a tabular form.}
\usage{
get.predinfo(x, conf_level = 0.9)
}
\arguments{
\item{x}{the output of the prediction function, of class sptotalPredOut}

\item{conf_level}{is the confidence level for a normal-based
confidence interval (default = 0.90)}
}
\value{
a list of tables with summary information about the predictions
}
\description{
Creates a table that shows the prediction, standard error, and
confidence interval for the prediction, as well as some summary information
about the sample.
}
