% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.predinfo.R
\name{get.predinfo}
\alias{get.predinfo}
\title{Display basic summary information in a tabular form.}
\usage{
get.predinfo(x, conf_level = 0.9)
}
\arguments{
\item{x}{the output of the \code{\link{predict.slmfit}()} function, of class \code{sptotalPredOut}}

\item{conf_level}{is the confidence level for a normal-based
confidence interval (default = 0.90).}
}
\value{
a list of three tables, including \itemize{
\item \code{simptab}, which contains the prediction and its standard error,
\item \code{confbounds}, which contains a confidence interval for the prediction, and
\item \code{outptmat}, a table of sampling information, including the number of sites sampled, the total number of sites, the total observed response, and the
observed average density (equal to the average response if all site areas are
equal).
}
}
\description{
Creates a list of tables that shows the prediction, standard error, and
confidence interval for the prediction, as well as some summary information
about the sample.
}
\examples{
data(exampledataset) ## load a toy data set
slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
predobj <- predict(slmobj)
get.predinfo(predobj)
}
