% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.slmfit.R
\name{logLik.slmfit}
\alias{logLik.slmfit}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{slmfit}(object, ...)
}
\arguments{
\item{object}{is a slmfit object}

\item{...}{additional argurments}
}
\value{
the log-likelihood of the slmfit object
}
\description{
Extract Log-Likelihood
}
