% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logLik.slmfit.R
\name{loglik.slmfit}
\alias{loglik.slmfit}
\title{Extract Log-Likelihood from a fitted \code{slmfit()} object}
\usage{
loglik.slmfit(object, ...)
}
\arguments{
\item{object}{is a slmfit object}

\item{...}{additional argurments}
}
\value{
the log-likelihood of the slmfit object
}
\description{
Extract Log-Likelihood from a fitted \code{slmfit()} object
}
\examples{
data(exampledataset) ## load a toy data set
slmobj <- slmfit(formula = counts ~ pred1 + pred2, data = exampledataset,
xcoordcol = 'xcoords', ycoordcol = 'ycoords', areacol = 'areavar')
loglik.slmfit(slmobj)
}
