% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.slmfit.R
\name{print.summary.slmfit}
\alias{print.summary.slmfit}
\title{Prints the summary of a fitted spatial linear model}
\usage{
\method{print}{summary.slmfit}(x, digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"), ...)
}
\arguments{
\item{x}{is an summary object generated from}

\item{digits}{is the number of digits to be displayed in the
model output}

\item{signif.stars}{is an option to show which predictors
are significant.
 \code{summary.slmfit}.}
}
\description{
Prints the summary of a fitted spatial linear model
}
