% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sptotal.R
\docType{package}
\name{sptotal}
\alias{sptotal}
\alias{sptotal-package}
\title{sptotal: A package used for performing finite population block
kriging on polygonal count data.}
\description{
The package provides an option to perform FPBK on counts assuming
perfect detection.
}
\details{
\code{sptotal} Main Functions:

\code{FPBKpred} is the function used to generate the kriged predictions
on the unsampled sites.

\code{FPBKoutput} takes the resulting object from FPBKpred to construct
confidence intervals, variograms, and maps.
}
