% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.slmfit.R
\name{summary.slmfit}
\alias{summary.slmfit}
\title{Summarizes a fitted spatial linear model}
\usage{
\method{summary}{slmfit}(object)
}
\arguments{
\item{object}{is an object generated from \code{slmfit}.}
}
\description{
Summarizes a fitted spatial linear model
}
